@@ -1,6 +1,6 @@
 {
 	"name": "PockyBot",
-	"version": "1.4.0",
+	"version": "1.6.0",
 	"description": "Spark bot that handles team recognition",
 	"main": "server.js",
 	"scripts": {
@@ -24,28 +24,27 @@
 	"author": "blake.mccririck@globalx.com.au",
 	"license": "ISC",
 	"dependencies": {
-		"@google-cloud/storage": "^2.3.4",
-		"ciscospark": "^1.48.0",
-		"express": "^4.16.4",
+		"@google-cloud/storage": "^2.5.0",
+		"ciscospark": "^1.59.0",
+		"express": "^4.17.1",
 		"libxmljs": "^0.19.5",
-		"pg": "^7.7.1",
+		"pg": "^7.11.0",
 		"pjson": "^1.0.9",
-		"rewiremock": "^3.12.3",
-		"string-width": "^2.1.1",
+		"rewiremock": "^3.13.7",
+		"string-width": "^4.1.0",
 		"unescape": "^1.0.1"
 	},
 	"devDependencies": {
-		"@types/express": "^4.16.0",
-		"@types/google-cloud__storage": "^1.7.2",
-		"@types/jasmine": "^3.3.5",
-		"@types/libxmljs": "^0.18.2",
-		"@types/node": "^10.12.18",
-		"@types/pg": "^7.4.11",
-		"@types/sinon": "^7.0.3",
-		"jasmine": "^3.3.1",
+		"@types/express": "^4.17.0",
+		"@types/jasmine": "^3.3.13",
+		"@types/libxmljs": "^0.18.3",
+		"@types/node": "^12.0.7",
+		"@types/pg": "^7.4.14",
+		"@types/sinon": "^7.0.12",
+		"jasmine": "^3.4.0",
 		"jasmine-ts": "^0.3.0",
-		"sinon": "^7.2.2",
+		"sinon": "^7.3.2",
 		"ts-node": "^7.0.1",
-		"typescript": "^2.9.2"
+		"typescript": "^3.5.1"
 	}
 }
