@@ -6,6 +6,8 @@ import { PegGiven, ResultRow, Role } from '../../models/database';
 import { DbUsers } from './db-interfaces';
 import { PockyDB as PockyDbInterface } from './db-interfaces';
 import QueryHandler from './query-handler-interface';
+import Utilities from '../utilities';
+import { PegGivenData } from '../../models/peg-given-data';
 
 export default class PockyDB implements PockyDbInterface {
 	private readonly sqlGivePegWithComment : string;
@@ -18,10 +20,13 @@ export default class PockyDB implements PockyDbInterface {
 	private config : Config;
 	private queryHandler : QueryHandler;
 	private dbUsers : DbUsers;
+	private readonly utilities : Utilities;
 
-	constructor(queryHandler : QueryHandler, dbUsers : DbUsers) {
+
+	constructor(queryHandler : QueryHandler, dbUsers : DbUsers, utilities : Utilities) {
 		this.queryHandler = queryHandler;
 		this.dbUsers = dbUsers;
+		this.utilities = utilities;
 
 		this.sqlGivePegWithComment = this.queryHandler.readFile('../../../database/queries/give_peg_with_comment.sql');
 		this.sqlPegsGiven = this.queryHandler.readFile('../../../database/queries/pegs_given.sql');
@@ -41,6 +46,7 @@ export default class PockyDB implements PockyDbInterface {
 	 *         2 on error
 	 */
 	async givePegWithComment(comment : string, receiver : string, sender = 'default_user') : Promise<number> {
+		// Check or create both sender and receiver users
 		try {
 			await Promise.all([this.dbUsers.existsOrCanBeCreated(sender), this.dbUsers.existsOrCanBeCreated(receiver)]);
 		} catch (error) {
@@ -49,7 +55,7 @@ export default class PockyDB implements PockyDbInterface {
 
 		let senderHasPegs : boolean;
 		try {
-			senderHasPegs = await this.hasSparePegs(sender);
+			senderHasPegs = await this.senderCanPeg(sender, comment);
 		} catch (error) {
 			return dbConstants.pegError;
 		}
@@ -74,25 +80,38 @@ export default class PockyDB implements PockyDbInterface {
 		}
 	}
 
-	async countPegsGiven(user : string) : Promise<number> {
+	async countPegsGiven(user : string, keywords : string[], penaltyKeywords : string[]) : Promise<number> {
 		let query : QueryConfig = {
 			name: 'pegsGiven',
 			text: this.sqlPegsGiven,
 			values: [user]
 		};
 
+		let givenPegs : PegGivenData[];
+
 		try {
-			let data = await this.queryHandler.executeQuery(query);
-			return data[0]['count'];
+			givenPegs = await this.queryHandler.executeQuery(query);
 		} catch (error) {
 			__logger.error(`[PockyDb.countPegsGiven] Error executing query to count pegs given by user ${user}`);
 			throw error;
 		}
+
+		const nonPenaltyPegs = this.utilities.getNonPenaltyPegs(givenPegs, keywords, penaltyKeywords);
+
+		return nonPenaltyPegs.length;
 	}
 
-	async hasSparePegs(user : string) : Promise<boolean> {
-		let count = this.countPegsGiven(user);
-		__logger.debug(`[PockyDb.hasSparePegs] Checking if user ${user} has spare pegs`);
+	async senderCanPeg(user : string, comment : string) : Promise<boolean> {
+		const keywords = this.config.getStringConfig('keyword');
+		const penaltyKeywords = this.config.getStringConfig('penaltyKeyword');
+
+		let count = this.countPegsGiven(user, keywords, penaltyKeywords);
+
+		__logger.debug(`[PockyDb.senderCanPeg] Checking if user ${user} has spare pegs`);
+
+		if (this.utilities.commentIsPenalty(comment, keywords, penaltyKeywords)) {
+			return true;
+		}
 
 		if (user === 'default_user' || this.config.checkRole(user, Role.Unmetered)) {
 			return true;
