@@ -2,10 +2,10 @@ import Trigger from '../../models/trigger';
 import Config from '../config';
 import constants from '../../constants';
 import { MessageObject } from 'ciscospark/env';
+import { Command } from '../../models/command';
 
 export default class Keywords extends Trigger {
-	readonly commandText : string = 'keywords';
-	readonly keywordsCommand : string = `(?: )*${this.commandText}(?: )*`;
+	readonly keywordsCommand : string = `(?: )*${Command.Keywords}(?: )*`;
 
 	config : Config;
 
@@ -21,7 +21,7 @@ export default class Keywords extends Trigger {
 	}
 
 	isToTriggerOnPM(message : MessageObject) : boolean {
-		return message.text.toLowerCase().trim() === this.commandText;
+		return message.text.toLowerCase().trim() === Command.Keywords;
 	}
 
 	async createMessage() : Promise<MessageObject> {
@@ -31,8 +31,15 @@ export default class Keywords extends Trigger {
 			newMessage += `* ${item}\n`;
 		});
 
+		const penaltyKeywords = this.config.getStringConfig('penaltyKeyword');
+		if (penaltyKeywords.length > 0) {
+			newMessage += '\n## Here is the list of keywords that can be used to apply a penalty to the sender\n\n';
+			newMessage += 'Penalty keywords do not count against the peg limit, and are *not* applied to messages that also include standard keywords.\n\n';
+			newMessage += penaltyKeywords.map(keyword => `* ${keyword}\n`);
+		}
+
 		return {
-				markdown: newMessage
+			markdown: newMessage
 		};
 	}
 }
