@@ -13,6 +13,7 @@ import Welcome from './welcome';
 import Help from './help';
 import Ping from './ping';
 import Keywords from './keywords';
+import Rotation from './rotation';
 import Default from './default';
 
 // Services
@@ -26,24 +27,27 @@ import QueryHandler from '../database/query-handler';
 import PockyDB from '../database/pocky-db';
 import DbUsers from '../database/db-users';
 import DbConfig from '../database/db-config';
+import Utilities from '../utilities';
 
 
 // Service instantiation
+const utilities = new Utilities();
 const queryHandler = new QueryHandler(new Client());
 const dbConfig = new DbConfig(queryHandler);
 const dbUsers = new DbUsers(spark, queryHandler);
-const database = new PockyDB(queryHandler, dbUsers);
+const database = new PockyDB(queryHandler, dbUsers, utilities);
 const config = new configService(dbConfig);
 
 database.loadConfig(config);
 config.updateAll();
 
 // Trigger instantiation
-const status = new Status(spark, database, config);
+const status = new Status(spark, database, config, utilities);
 const welcome = new Welcome(config);
 const keywords = new Keywords(config);
 const help = new Help(config);
 const ping = new Ping();
+const rotation = new Rotation(config);
 const defaultTrigger = new Default();
 
 const triggers : Trigger[] = [
@@ -52,6 +56,7 @@ const triggers : Trigger[] = [
 	ping,
 	status,
 	keywords,
+	rotation,
 	defaultTrigger
 ];
 
